Convert the 1x high bit block + 7x 7 bit messages format into 8 bits.
Korg data dump packs 8 bit data values into a stream of 7 bytes. This
work decodes that data stream.

In the format, a lead block holds the 8th bit of the next 7 bytes. I
call this the "high" byte. To convert into an array of normal bytes,
read all non-high bytes, and lookup & read in the highest bit from
the preceeding high byte.